Modification des erreur dans le cpp06

+++ cpp06/ex01/srcs/Serializer.cpp
@@ -1,4 +1,5 @@
 #include "../includes/Serializer.hpp"
+#include <cstdint>
 
 Serializer::Serializer()
 {
@@ -21,7 +22,7 @@ Serializer::~Serializer()
 
 uintptr_t	Serializer::serialize(Data *ptr)
 {
-	return reinterpret_cast<unsigned long>(ptr);
+	return reinterpret_cast<uintptr_t>(ptr);
 }
 
 Data*	Serializer::deserialize(uintptr_t raw)
